Fix datatype

+++ src/main/java/com/nastya/images/dao/SocialNetworkDao.java
@@ -3,6 +3,6 @@ package com.nastya.images.dao;
 import com.nastya.images.entity.SocialNetwork;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SocialNetworkDao extends JpaRepository<SocialNetwork, Short> {
+public interface SocialNetworkDao extends JpaRepository<SocialNetwork, Integer> {
 
 }
