+++ src/main/java/com/nastya/images/dao/UserDao.java
@@ -3,7 +3,7 @@ package com.nastya.images.dao;
 import com.nastya.images.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDao extends JpaRepository<User, Short> {
+public interface UserDao extends JpaRepository<User, Integer> {
 
     User findByLogin(String login);
 }
